requirements.txt
deep-translator==1.11.4
faiss-gpu==1.7.2
torch==2.3.0
torchvision==0.18.0
translate==3.6.1
gdown==5.2.0
elasticsearch==7.17.1
open_clip_torch==2.26.1
unicorn==2.1.0
fastapi==0.115.0
pyvi==0.1.1
langdetect==1.0.9
underthesea==6.8.4
transformers==4.44.2
matplotlib==3.9.2
sentence_transformers==3.1.1
git+https://github.com/openai/CLIP.git
sentencepiece==0.2.0